requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
langchain
openai
pydantic
